конструктор FileBackedTaskManager использует Managers.getDefaultHistory(); создан класс TaskUtils для разгрузки FileBackedTaskManager; убрано создание и удаление файла, теперь save() перезаписывает актуальный файл.
