+++ src/manager/TaskManager.java
@@ -7,8 +7,6 @@ import java.util.List;
 
 public interface TaskManager {
 
-    int generatorId();
-
     void addTask(Task task);
 
     void addSubTask(SubTask subTask);
